refactor: clean up unused PDF worker implementations
- Created worker-types.ts to centralize type definitions
- Updated imports to use the new types file
- Deleted unused worker implementations:
  - pdf-worker.ts (only used for types, now replaced)
  - pdf-worker-optimized.ts (completely unused)
  - pdf-worker-simplified.ts (completely unused)
- Fixed TRIGGER_THRESHOLD constant name issue
- Kept only pdf-worker-with-fonts.ts as the actual implementation

This removes ~1000+ lines of unused code that was creating confusion.
The actual worker (pdf-worker-with-fonts.ts) is compiled to public/pdf-worker.js.
